+++ src/main.lua
@@ -26,24 +26,37 @@ chalk = mods["SGG_Modding-Chalk"]
 ---@module 'SGG_Modding-ReLoad'
 reload = mods['SGG_Modding-ReLoad']
 
+
+---@module 'SGG_Modding-SJSON'
+sjson = mods['SGG_Modding-SJSON']
+
 ---@module 'adamant-config-Modpack'
 config = chalk.auto('config.lua')
 public.config = config
 
-
+local function ApplyModdedMark()
+    local file = rom.path.combine(rom.paths.Content, 'Game/Text/en/HelpText.en.sjson')
+    sjson.hook(file, function(data)
+        for _, v in ipairs(data.Texts) do
+            if v.Id == 'UI_RoomCount' then
+                v.DisplayName = v.DisplayName .. "\n\n\n\n\nModded Run"
+                break
+            end
+        end
+    end)
+end
 local function on_ready()
     import_as_fallback(rom.game)
 
+    ApplyModdedMark()
     import("def.lua")
     import("logic.lua")
 end
 
 local function on_reload()
     import_as_fallback(rom.game)
-    -- import("logic.lua")
     import("def.lua")
-    
-    import ("ui.lua")
+    import("ui.lua")
 end
 
 local loader = reload.auto_single()